Bump patch version
This is deemed a bug-fix, even though
it changes behaviour that was tested,
since that behaviour was based on a
misunderstanding of the format. PSS/E
documents that names are a maximum
length of 12 characters long, but
evidently if leading and trailing
whitespace is included then there are
files with names >12 characters long.
Assuming these files are valid, having
been produced by the PSS/E software,
then it seems leading and trailing
whitespace should not be included in
the names (only mid-name whiespace).
See #44.